TQ: Begin integration with sled-agent (#9465)
This PR starts integrating trust quorum with sled-agent. It does this by
starting the `trust_quorum::NodeTask` as a long running task. It also
updates the peer status of this task by enhancing the ddmd poller in a
manner similar to LRTQ.

While the task itself runs on each sled and sets up sprockets
connections among peers, it is largely inert, as RSS will never create
an initial configuration, and no state will be persisted. This behavior
is currently behind a constant `TRUST_QUORUM_INTEGRATION_ENABLED`, which
will be removed from the code base when we want to start using trust
quorum and are capable of upgrading from LRTQ. We plan to use this gate
in a few other places as we continue to integrate trust quorum into the
rest of the control plane.

It was also necessary to recognize that the certificates for sprockets
connections match the trust quorum nodes on each simulated sled. This is
done by ensuring that the `BaseboardId` matches what's in the
certificate. a4x2 will inject SMBIOS Type 1 table information, which is
read by `sled-hardware` when a PC is detected as the running platform.
This injection mechanism required changes to propolis and falcon which
have now merged. An a4x2 PR is coming soon to utilize these changes and
complete the injection mechanism. However it's preferred to merge this
PR first since trust quorum is disabled, and because further changes in
a4x2 want to use the `sp-sim` config changes in this PR.