[omdb] exhume `omdb db saga show` from #4378
In the days of yore, @jmpesp wrote an `omdb db saga show` command in PR
#4378, which prints a trace of the execution of a given saga.
Unfortunately, this code never made it to `main`. Some parts of #4378
adding _other_ saga-related OMDB commands merged in separate PRs, but
the `saga show` command never made it in. Instead, when it's necessary
to display the execution history of a saga for debugging, folks have
maintained separate `omdb-sagas` binaries built from #4378 and copied
them onto the dogfood rack or racklettes as needed for debugging.

This is not an ideal situation. If an issue occurs on a customer rack
rather than in development, it would be useful for support to be able to
display these saga execution traces without copying a random binary onto
the customer's rack, which may not be permitted depending on the
customer environment. And, when the subcommand is not merged to the
`main` branch and is instead maintained separately, it may drift
substantially from the rest of the codebase, potentially encountering
schema incompatibilities in future versions.

Therefore, this PR digs up the `omdb db saga show` command from #4378
and does the minimum amount of changes necessary to make it compile with
the current `main`. Almost all of the added code was written by @jmpesp,
and I've just copied it and pasted it, and fixed compiler errors and
such. While I'm open to addressing some of the suggestions from @dap's
[review of #4378][1], I felt like the priority was to just get the code
pulled forward into a working state.

[1]:
https://github.com/oxidecomputer/omicron/pull/4378#pullrequestreview-1963841208

Authored-by: James MacMahon <james@oxide.computer> Co-authored-by: Eliza
Weisman <eliza@elizas.website>