[nexus] refactor SP ereport processing into `nexus-types`
This commit moves the code that processes raw JSON representations of SP
ereports from the `ereport_ingester` background task to
`nexus_types::fm::ereport`. While a production system will only perform
this processing when an ereport is received from a service processor via
MGS in the `ereprot_ingester` task, I'd like to be able to also invoke
the same logic with hard-coded example JSON ereports for test purposes.
Currently, the test code for diagnosis engines in #9346 implemented its
own version of this which was slightly different from the
ereport_ingester version, so I've moved it to here so that we can use
the same logic when processing test inputs.