[nexus] refactor SP ereport processing into `nexus-types` (#9480)
This commit moves the code that processes raw JSON representations of SP
ereports from the `ereport_ingester` background task to
`nexus_types::fm::ereport`. While a production system will only perform
this processing when an ereport is received from a service processor via
MGS in the `ereprot_ingester` task, I'd like to be able to also invoke
the same logic with hard-coded example JSON ereports for test purposes.
Currently, the test code for diagnosis engines in #9346 implemented its
own version of this which was slightly different from the
ereport_ingester version, so I've moved it to here so that we can use
the same logic when processing test inputs.

Note that this change was cherry-picked from #9346, where I had actually
written tests that needed to use this logic. I figured it was a small
and uncontroversial enough refactor (that makes no functional change
whatsoever) to pull it out to merge now, in an attempt to keep that
branch from getting even bigger and touching even more files.