Add support for dual-stack external addressing (#9470)
- Adds enum for tracking the external IP addresses for an instance or
service, which includes the three fields that were previously separate:
SNAT, ephemeral, and list of floating addresses.
- Makes the previous source NAT config type IP address specific, with a
"generic" variant for services. This is a bit lazy, but we can lift the
IP version out of the contained IP address and into the source NAT type
itself in the future if we need to.
- Add new version of the sled-agent API with these new types, plus
conversions with the previous version
- Completely handle dual-stack OPTE ports, closes #9247
- Closes #9318