Add `parent_blueprint` to `PlanningInput` (#9481)
There are two main changes in this PR:

* `PlanningInput` now has a `parent_blueprint` field
* The `Planner` no longer takes a separate parent blueprint argument
(since it comes with the planning input)

There are a bunch of other mechanical changes, mostly related to
trafficking in `Arc<Blueprint>`s in various places.

The rationale for change this is to ensure correctness as we get into
#5552; I wrote this up in the new doc comment on
`PlanningInputFromDb::assemble()`; feedback welcome!

This is staged on top of #9473. I had tried to make this change before
doing that work, and updating the planner tests was a true nightmare.
Happy to report updating the tests after that change was quite easy!
(Didn't touch the planner tests at all, which only left a few that
constructed `PlanningInput`s outside the simulator.)