Add `Blueprint` methods for "info that ought to be configurable but isn't yet" (#9510)
We have three spots where the planner needs information that has to come
from the operator, but it's forced to infer it from the parent blueprint
instead of some operator-controllable database bits:

* Nexus's config (TLS / upstream DNS)
* Upstream NTP config
* The set of IPs on which we should run DNS

This PR is just a bit of reorganization: `Blueprint` now exposes methods
for all three of these instead of the planner scraping through the
blueprint itself. It also squashes a bit of duplication where
`DataStore` also needed one of these pieces of info too.

(This comes out of auditing our use of reading the expunged zones of a
blueprint in the context of #5552 - I'd like to minimize the uses of
`BlueprintZoneDisposition::any`, and this pulls three of them out into a
set of related methods together.)