[sled-agent-types] move functional code to impls module (#9514)
In #9488 we moved versioned Sled Agent types to the
sled-agent-types-versions crate. In that PR we moved functional code
into that crate as well, living in the same files as types. However,
dealing with updating this code can easily go wrong, particularly around
merge conflicts.

Move this kind of functional code to an `impls` module, so that it is
always implemented on the latest versions of types.