[nexus] add cases to FM sitreps (#9492)
Cases, [as described in RFD 603][1], give diagnosis engines a mechanism
to group together ereports (and eventually, other fault management data)
as part of a sitrep. This branch adds database tables for representing
cases, and for associating ereports with cases. Subsequent branches will
add a mechanism for sitreps to request alerts, which will also be
associated with cases.

This branch was factored out from #9346, where I've been prototyping a
diagnosis engine implementation. Although aspects of how the DEs
themselves will work are still in flux, I think that cases, along with
other parts of the database machinery such as the impending
alert-requests branch, are pretty much nailed down.

[1]: https://rfd.shared.oxide.computer/rfd/0603#_cases