Separate transit IPs by version in the database (#9542)
This is a follow-up to #9389 that I missed. We were already separating
transit IPs by version in the code before and after the database. This
pushes that separation all the way into the database, which will make it
easier to implement updates to dual-stack NICs in the public API.
Separating them lets clients update the transit IPs for only one
version, without some janky read-modify-write code that ensures we don't
clobber the transit IPs for the _other_ version.