Local storage [4/4]: Use raw zvols
Request the new *raw* zvol type when creating disks backed by local
storage. Using raw zvols requires waiting for them to be initialized,
which the instance start saga is now responsible for polling.

In the instance start saga, ensure each raw zvol in parallel because
this can take some time.

This PR also explicitly removes the inherited encryption for these
zvols. This is accomplished by changing `EncryptionDetails` to be an
enum where the call site now has to explicitly choose between inheriting
the parent dataset's encryption, using `aes-256-gcm`, and explicitly
setting encryption off.

Also chopped out is setting `volblocksize`: this was a mistake to set,
and it should instead be left to the default value.

Still to do is to revisit the overhead reserved for zvols: the previous
emperical value was not for the new raw zvol type. This is tracked by
oxidecomputer/omicron#9591.

Fixes #9519.