Blippy: Fix checks now that blueprints have sled subnets
Blippy has had checks on sled subnet consistency for a while, but they
were implemented by inferring each sled's subnet from its zones' IPs.
We added explicit sled subnets to blueprints in #9416; this changes
blippy's sled-subnet-related checks to use that new field.

(This is work that fell out of a larger PR that isn't ready yet, and I'm
opening it separately to reduce diff clutter in that PR.)