Add missing newlines in api docstrings.
At the moment, some docstrings in the generated docs lack newlines, and string
together multiple sentences without punctuation. For example, the `provisioned`
field in the `silo_utilization_list` method is a bit garbled:
https://docs.oxide.computer/api/silo_utilization_list.

> Accounts for resources allocated by in silos like CPU or memory for running
> instances and storage for disks and snapshots Note that CPU and memory
> resources associated with a stopped instances are not counted here

This patch adds the missing newlines and punctuation to make this docstring
readable. Note that these changes won't show up in the generated docs until we
cut a new release.

Note: if this change makes sense, I'll also ask Claude to review all docstrings
in this crate for similar issues.