TQ: don't claim tasks panicked when they haven't
`tokio::task::JoinSet::join_next_with_id` can return a `JoinError` if a
task has panicked _or_ if a task sapwned on the `JoinSet` was aborted
through its `AbortHandle`...which the trust-quorum `ConnectionManager`
will do when disconnecting a client connection:
https://github.com/oxidecomputer/omicron/blob/e00aabaee707ae1fab7203d72c0243f8dacf5403/trust-quorum/src/connection_manager.rs#L866-L886

This is a normal operation and probably deserves to be logged at a less
verbose level than `Warn`, and the log line which suggests both
cancelled _and_ panicked `JoinError`s are panics is incorrect. This is a
potential red herring while debugging, so I've changed it to be more
accurate.