Add measurements to sled-agent
This introduces a `MeasurementsHandle` function which can
be used to access the current set of reference meaurements.
Right now, this only includes what's in the install dataset
but future work will include getting measurements from
the artifact depot.

This shuffles some things around in the early bootup to allow
reading `OmicronSledConfig` via the ledger task at cold boot.
Measurements are needed early in sled-agent startup for use with
trust-quorum.