[3/n] [reconfigurator-sim] add an operation log (#9365)
Add an operation log to the simulator, as well as ways to undo, redo and
restore to the last operation. Modeled pretty directly after how
Jujutsu's operation log works.

The biggest change here is that the current pointer is now managed by
the simulator rather than reconfigurator-cli, so that it can be
versioned as part of the operation.