TQ: Plumb initial configuration through handoff
This ensures that when trust quorum is enabled the initial configuration
gets passed up through the handoff to Nexus and put in CRDB.

It was tested in a4x2 with `TRUST_QUORUM_INTEGRATION_ENABLED=true`, a
small [fix](https://github.com/oxidecomputer/testbed/pull/236) for a4x2,
and a fix in this PR to plumb through the correct sp-sim model number so
that handoff works correctly in a4x2.

This PR should have no impact on main, as
`TRUST_QUORUM_INTEGRATION_ENABLED` remains set to `false`.